deal with the fixed prefix for git clones too

+++ classes/package.php
@@ -48,13 +48,12 @@ class Package
 
 		foreach ($config['sources'] as $source)
 		{
-			$zips = array(
-				'http://' . rtrim($source, '/').'/fuel-'.$package.'/zipball/'.$version,
-				'http://' . rtrim($source, '/').'/'.$package.'/zipball/'.$version,
-			);
+			$packages = array('fuel-'.$package, $package);
 
-			foreach ($zips as $zip_url)
+			foreach ($packages as $package)
 			{
+				$zip_url = 'http://' . rtrim($source, '/').'/'.$package.'/zipball/'.$version;
+
 				if ($fp = @fopen($zip_url, 'r'))
 				{
 					// We don't actually need this, just checking the file is there
@@ -163,31 +162,38 @@ HELP;
 		$zip_file = $tmp_folder . '.zip';
 		@copy($zip_url, $zip_file);
 
-		$unzip = new \Unzip;
-		$files = $unzip->extract($zip_file, $tmp_folder);
-
-		// Grab the first folder out of it (we dont know what it's called)
-		list($tmp_package_folder) = glob($tmp_folder.'/*', GLOB_ONLYDIR);
-
-		$package_folder = PKGPATH . $package;
-
-		// Move that folder into the packages folder
-		rename($tmp_package_folder, $package_folder);
-
-		unlink($zip_file);
-		rmdir($tmp_folder);
-
-		foreach ($files as $file)
+		if (file_exists($zip_file))
 		{
-			$path = str_replace($tmp_package_folder, $package_folder, $file);
-			chmod($path, octdec(755));
-			\Cli::write("\t" . $path);
+			$unzip = new \Unzip;
+			$files = $unzip->extract($zip_file, $tmp_folder);
+
+			// Grab the first folder out of it (we dont know what it's called)
+			list($tmp_package_folder) = glob($tmp_folder.'/*', GLOB_ONLYDIR);
+
+			$package_folder = PKGPATH . $package;
+
+			// Move that folder into the packages folder
+			rename($tmp_package_folder, $package_folder);
+
+			unlink($zip_file);
+			rmdir($tmp_folder);
+
+			foreach ($files as $file)
+			{
+				$path = str_replace($tmp_package_folder, $package_folder, $file);
+				chmod($path, octdec(755));
+				\Cli::write("\t" . $path);
+			}
+		}
+		else
+		{
+			\Cli::write('Package could not be found', 'red');
 		}
 	}
 
 	public static function _clone_package_repo($source, $package, $version)
 	{
-		$repo_url = 'git://' . rtrim($source, '/').'/fuel-'.$package . '.git';
+		$repo_url = 'git://' . rtrim($source, '/').'/'.$package . '.git';
 
 		\Cli::write('Downloading package: ' . $repo_url);
